feat(multi-account): Implemented account switcher logic

+++ app/utils/redis.server.ts
@@ -16,8 +16,9 @@ const ensureClientOpen = () => {
 };
 
 const get = async (key: string) => {
+	if (!key) return console.trace("WARNING: Redis GET called without key");
 	await ensureClientOpen();
-	// if (cache[key] || cache[key] === null) return cache[key];
+	if (cache[key] || cache[key] === null) return cache[key];
 
 	const value = await client.get(key);
 	if (!value) {
@@ -30,15 +31,20 @@ const get = async (key: string) => {
 };
 
 const del = async (key: string) => {
+	if (!key) return console.trace("WARNING: DEL called without key");
 	await ensureClientOpen();
+
 	cache[key] = null;
 	return client.del(key);
 };
 
 const set = async (key: string, value: any) => {
+	if (!key) return console.trace("WARNING: SET called without key");
+	if (!value) return console.trace("WARNING: SET called without value");
+	const encoded = JSON.stringify(value);
 	await ensureClientOpen();
 	cache[key] = value;
-	return client.set(key, JSON.stringify(value));
+	return client.set(key, encoded || "");
 };
 
 export { get, set, del, client };
